requirements.txt
notion-client==2.2.1
plyer==2.1.0
python-dateutil==2.8.2
schedule==1.2.0
python-dotenv==1.0.0 